==> analysis/v1/4_exclusion_analysis.Rmd <==
---
title: "Kitty Act V1 Exclusion Analysis"
author: "Jaime Chen"
date: "2025-01-19"
output:
  pdf_document: default
  html_document: default
---
  
```{r setup, warning=FALSE, message=F}
library(here)
library(tidyverse)
library(ggplot2)
library(viridisLite)
library(lme4)
library(ggstance)
library(cowplot)
library(AICcmodavg)
library(RColorBrewer)
library(jsonlite)
library(tidyjson)
library(rlang)
library(car)
library(nnet)
library(mlogit)
library(lmerTest)
library(assertthat)
library(knitr)
source("helper.R")

data_path <- here("processed_data", "kitty-act-v1-alldata.csv")

sampling_data_path_v1 <- here("processed_data", "kitty-act-v1-processed_data_sampling_v1.csv")
sampling_data_path_v1_1 <- here("processed_data", "kitty-act-v1-processed_data_sampling_v1_1.csv")

test_data_path_v1 <- here("processed_data", "kitty-act-v1-processed_data_v1.csv")
test_data_path_v1_1 <- here("processed_data", "kitty-act-v1-processed_data_v1_1.csv")

test_data_long_path_v1 <- here("processed_data", "kitty-act-v1-processed_data_long_v1.csv")
test_data_long_path_v1_1 <- here("processed_data", "kitty-act-v1-processed_data_long_v1_1.csv")

participant_info_path <- here("processed_data","kitty-act-v1-participant_info.csv")
write_path <- here("processed_data")
figure_path <- here("figures")
```

# Data Processing

## Read in data

```{r, warning=FALSE, message=F}
v1_participants = c("p004", "p005", "p006", "p007", "p008", "p009", "p010", "p011", "p012", "p013", "p014", "p015", "p016", "p017", "p018", "p019", "p020", "p021", "p022", "p023", "p024", "p025", "p026")

d <- read_csv(data_path) %>%
  mutate(number_training_images_order = case_when(
    number_training_image_order_vector == "[1,3]" ~ "One-Three",
    number_training_image_order_vector == "[3,1]" ~ "Three-One",
  ))

d_v1 <- d %>% filter(subject_id %in% v1_participants)
d_v1_1 <- d %>% filter(! subject_id %in% v1_participants)

sampling_data_v1 <- read_csv(sampling_data_path_v1) %>%
  mutate(number_training_images_order = case_when(
    number_training_image_order_vector == "[1,3]" ~ "One-Three",
    number_training_image_order_vector == "[3,1]" ~ "Three-One",
  ))
sampling_data_v1_1 <- read_csv(sampling_data_path_v1_1) %>%
  mutate(number_training_images_order = case_when(
    number_training_image_order_vector == "[1,3]" ~ "One-Three",
    number_training_image_order_vector == "[3,1]" ~ "Three-One",
  ))

test_array_clean_v1 <- read_csv(test_data_path_v1) %>%
  mutate(number_training_images_order = case_when(
    number_training_image_order_vector == "[1,3]" ~ "One-Three",
    number_training_image_order_vector == "[3,1]" ~ "Three-One",
  ))
test_array_clean_v1_1 <- read_csv(test_data_path_v1_1) %>%
  mutate(number_training_images_order = case_when(
    number_training_image_order_vector == "[1,3]" ~ "One-Three",
    number_training_image_order_vector == "[3,1]" ~ "Three-One",
  ))

test_array_options_clean_v1 <- read_csv(test_data_long_path_v1) %>%
  mutate(number_training_images_order = case_when(
    number_training_image_order_vector == "[1,3]" ~ "One-Three",
    number_training_image_order_vector == "[3,1]" ~ "Three-One",
  ))
test_array_options_clean_v1_1 <- read_csv(test_data_long_path_v1_1) %>%
  mutate(number_training_images_order = case_when(
    number_training_image_order_vector == "[1,3]" ~ "One-Three",
    number_training_image_order_vector == "[3,1]" ~ "Three-One",
  ))

#participant info
participant_info <- read_csv(participant_info_path) %>%
  filter(pilot == "no")
```

## compute ages and integrate with main data

```{r}
participant_info <- participant_info %>%
  mutate(resid_month_decimal=age_day_count/30.417) %>%
  mutate(age_dec = age_mos+resid_month_decimal,
         age_years=age_mos/12) %>%
  mutate(
    age_years_c=age_years-mean(age_years)
  )

average_age_v1 <- participant_info %>% filter(subject_id %in% v1_participants) %>% 
  summarize(
    N=n(),
    mean_age=mean(age_years,na.rm=TRUE),
    sd_age=sd(age_years,na.rm=TRUE),
    min_age=min(age_years,na.rm=TRUE),
    max_age=max(age_years,na.rm=TRUE)
  )

average_age_v1_1 <- participant_info %>% filter(! subject_id %in% v1_participants) %>% 
  summarize(
    N=n(),
    mean_age=mean(age_years,na.rm=TRUE),
    sd_age=sd(age_years,na.rm=TRUE),
    min_age=min(age_years,na.rm=TRUE),
    max_age=max(age_years,na.rm=TRUE)
  )

average_age_v1 %>%
  kable()

average_age_v1_1 %>%
  kable()

participant_info_ages_v1 <- participant_info %>% 
  filter(subject_id %in% v1_participants) %>% 
  select(subject_id,age_group,age_dec,age_years,age_years_c)

participant_info_ages_v1_1 <- participant_info %>% 
  filter(! subject_id %in% v1_participants) %>% 
  select(subject_id,age_group,age_dec,age_years,age_years_c)

d_v1 <- merge(d_v1, participant_info_ages_v1)
sampling_data_v1 <- merge(sampling_data_v1, participant_info_ages_v1)
test_array_clean_v1 <- merge(test_array_clean_v1, participant_info_ages_v1)
test_array_options_clean_v1 <- merge(test_array_options_clean_v1, participant_info_ages_v1)

d_v1_1 <- merge(d_v1_1, participant_info_ages_v1_1)
sampling_data_v1_1 <- merge(sampling_data_v1_1, participant_info_ages_v1_1)
test_array_clean_v1_1 <- merge(test_array_clean_v1_1, participant_info_ages_v1_1)
test_array_options_clean_v1_1 <- merge(test_array_options_clean_v1_1, participant_info_ages_v1_1)
```

## Summarize Test Data

```{r}
## shorter test representation
subj_test_accuracy_v1 <- test_array_clean_v1 %>%
  group_by(participant,number_training_images,round,trial_counter,trial_type,current_category_training_level,current_category_label_level,current_category_kind,final_choice_array,total_number_correct_options) %>%
  summarize(
    total_choices = n(),
    ground_truth_correct_choices = sum(test_choice_type_label_consistent),
    ground_truth_incorrect_choices = sum(test_choice_type_label_consistent==0),
    training_consistent_choices = sum(test_choice_type_training_consistent),
    training_inconsistent_choices = sum(test_choice_type_training_consistent==0),
    training_category_match_subordinate_choices = sum(test_choice_match_category==1 & test_choice_type == "subordinate"),
    training_category_match_basic_choices = sum(test_choice_match_category==1 & test_choice_type == "basic"),
    training_category_match_superordinate_choices = sum(test_choice_match_category==1 & test_choice_type == "superordinate"),
    training_category_match_subordinate_percent = training_category_match_subordinate_choices,
    training_category_match_basic_percent = training_category_match_basic_choices,
    training_category_match_superordinate_percent = training_category_match_superordinate_choices/2
  )

subj_test_accuracy_v1_1 <- test_array_clean_v1_1 %>%
  group_by(participant,number_training_images,round,trial_counter,trial_type,current_category_training_level,current_category_label_level,current_category_kind,final_choice_array,total_number_correct_options) %>%
  summarize(
    total_choices = n(),
    ground_truth_correct_choices = sum(test_choice_type_label_consistent),
    ground_truth_incorrect_choices = sum(test_choice_type_label_consistent==0),
    training_consistent_choices = sum(test_choice_type_training_consistent),
    training_inconsistent_choices = sum(test_choice_type_training_consistent==0),
    training_category_match_subordinate_choices = sum(test_choice_match_category==1 & test_choice_type == "subordinate"),
    training_category_match_basic_choices = sum(test_choice_match_category==1 & test_choice_type == "basic"),
    training_category_match_superordinate_choices = sum(test_choice_match_category==1 & test_choice_type == "superordinate"),
    training_category_match_subordinate_percent = training_category_match_subordinate_choices,
    training_category_match_basic_percent = training_category_match_basic_choices,
    training_category_match_superordinate_percent = training_category_match_superordinate_choices/2
  )

subj_test_accuracy_long_v1 <- subj_test_accuracy_v1 %>%
  select(participant,number_training_images,round,trial_counter,trial_type,current_category_training_level,current_category_label_level,current_category_kind,final_choice_array,total_number_correct_options,training_category_match_subordinate_percent,training_category_match_basic_percent,training_category_match_superordinate_percent) %>%
  group_by(participant,number_training_images,round,trial_counter,trial_type,current_category_training_level,current_category_label_level,current_category_kind,final_choice_array,total_number_correct_options) %>%
  pivot_longer(cols=training_category_match_subordinate_percent:training_category_match_superordinate_percent,names_to = "choice_type",values_to = "percent_chosen")

subj_test_accuracy_long_v1_1 <- subj_test_accuracy_v1_1 %>%
  select(participant,number_training_images,round,trial_counter,trial_type,current_category_training_level,current_category_label_level,current_category_kind,final_choice_array,total_number_correct_options,training_category_match_subordinate_percent,training_category_match_basic_percent,training_category_match_superordinate_percent) %>%
  group_by(participant,number_training_images,round,trial_counter,trial_type,current_category_training_level,current_category_label_level,current_category_kind,final_choice_array,total_number_correct_options) %>%
  pivot_longer(cols=training_category_match_subordinate_percent:training_category_match_superordinate_percent,names_to = "choice_type",values_to = "percent_chosen")

overall_accuracy_by_condition_v1 <- subj_test_accuracy_long_v1 %>%
  ungroup() %>%
  group_by(number_training_images,current_category_training_level,choice_type) %>%
  summarize(
    N=n(),
    average_percent=mean(percent_chosen)
  ) %>%
  mutate(
    choice_type = str_remove(choice_type,"training_category_match_"),
    choice_type = str_remove(choice_type,"_percent")
  )

overall_accuracy_by_condition_v1_1 <- subj_test_accuracy_long_v1_1 %>%
  ungroup() %>%
  group_by(number_training_images,current_category_training_level,choice_type) %>%
  summarize(
    N=n(),
    average_percent=mean(percent_chosen)
  ) %>%
  mutate(
    choice_type = str_remove(choice_type,"training_category_match_"),
    choice_type = str_remove(choice_type,"_percent")
  )

# looking at if 1- or 3-Item first has an effect

overall_accuracy_by_condition_round_v1 <- subj_test_accuracy_long_v1 %>%
  ungroup() %>%
  group_by(round,number_training_images,current_category_training_level,choice_type) %>%
  summarize(
    N=n(),
    average_percent=mean(percent_chosen)
  ) %>%
  mutate(
    choice_type = str_remove(choice_type,"training_category_match_"),
    choice_type = str_remove(choice_type,"_percent")
  )

overall_accuracy_by_condition_round_v1_1 <- subj_test_accuracy_long_v1_1 %>%
  ungroup() %>%
  group_by(round,number_training_images,current_category_training_level,choice_type) %>%
  summarize(
    N=n(),
    average_percent=mean(percent_chosen)
  ) %>%
  mutate(
    choice_type = str_remove(choice_type,"training_category_match_"),
    choice_type = str_remove(choice_type,"_percent")
  )

# accuracy based on training condition (N, I, B) and ground-truth (same level or 1 level up)

overall_accuracy_label_level_by_condition_v1 <- subj_test_accuracy_long_v1 %>%
  ungroup() %>%
  group_by(number_training_images,current_category_training_level,current_category_label_level,choice_type) %>%
  summarize(
    N=n(),
    average_percent=mean(percent_chosen)
  ) %>%
  mutate(
    choice_type = str_remove(choice_type,"training_category_match_"),
    choice_type = str_remove(choice_type,"_percent")
  )

overall_accuracy_label_level_by_condition_v1_1 <- subj_test_accuracy_long_v1_1 %>%
  ungroup() %>%
  group_by(number_training_images,current_category_training_level,current_category_label_level,choice_type) %>%
  summarize(
    N=n(),
    average_percent=mean(percent_chosen)
  ) %>%
  mutate(
    choice_type = str_remove(choice_type,"training_category_match_"),
    choice_type = str_remove(choice_type,"_percent")
  )

# again, looking at if 1- or 3-Item first has an effect on above accuracy

overall_accuracy_label_level_by_condition_round_v1 <- subj_test_accuracy_long_v1 %>%
  ungroup() %>%
  group_by(round,number_training_images,current_category_training_level,current_category_label_level,choice_type) %>%
  summarize(
    N=n(),
    average_percent=mean(percent_chosen)
  ) %>%
  mutate(
    choice_type = str_remove(choice_type,"training_category_match_"),
    choice_type = str_remove(choice_type,"_percent")
  )

overall_accuracy_label_level_by_condition_round_v1_1 <- subj_test_accuracy_long_v1_1 %>%
  ungroup() %>%
  group_by(round,number_training_images,current_category_training_level,current_category_label_level,choice_type) %>%
  summarize(
    N=n(),
    average_percent=mean(percent_chosen)
  ) %>%
  mutate(
    choice_type = str_remove(choice_type,"training_category_match_"),
    choice_type = str_remove(choice_type,"_percent")
  )


### longer test representation
subj_test_accuracy_all_v1 <- test_array_options_clean_v1 %>%
  group_by(participant,number_training_images,round,trial_counter,trial_type,current_category_training_level,current_category_label_level,current_category_kind,final_choice_array,total_number_correct_options_label,total_number_correct_options_training) %>%
  summarize(
    accuracy_training = mean(is_match_to_training, na.rm=TRUE),
    accuracy_label = mean(is_match_to_label,na.rm=TRUE),
    hit_rate_training = mean(hit_training,na.rm=TRUE),
    hit_rate_label = mean(hit_label,na.rm=TRUE),
    false_alarm_rate_training = mean(false_alarm_training,na.rm=TRUE),
    false_alarm_rate_label = mean(false_alarm_label,na.rm=TRUE)
  ) %>%
  mutate(
    hit_rate_training_adj= case_when(
      hit_rate_training==1 ~ 1 - 1/(2*total_number_correct_options_training),
      hit_rate_training==0 ~ 1/(2*total_number_correct_options_training),
      TRUE ~ hit_rate_training
    ),
    hit_rate_label_adj= case_when(
      hit_rate_label==1 ~ 1 - 1/(2*total_number_correct_options_label),
      hit_rate_label==0 ~ 1/(2*total_number_correct_options_label),
      TRUE ~ hit_rate_label
    ),
    false_alarm_rate_training_adj= case_when(
      false_alarm_rate_training==0 ~ 1/(2*total_number_correct_options_training),
      false_alarm_rate_training==1 ~ 1 - 1/(2*total_number_correct_options_training),
      TRUE ~ false_alarm_rate_training
    ),
    false_alarm_rate_label_adj= case_when(
      false_alarm_rate_label==0 ~ 1/(2*total_number_correct_options_label),
      false_alarm_rate_label==1 ~ 1 - 1/(2*total_number_correct_options_label),
      TRUE ~ false_alarm_rate_label
    )
  ) %>%
  mutate(
    dprime_training=qnorm(hit_rate_training_adj) - qnorm(false_alarm_rate_training_adj),
    dprime_label=qnorm(hit_rate_label_adj) - qnorm(false_alarm_rate_label_adj),
    c_training=-.5*(qnorm(hit_rate_training_adj) + qnorm(false_alarm_rate_training_adj)),
    c_label=-.5*(qnorm(hit_rate_label_adj) + qnorm(false_alarm_rate_label_adj)))

subj_test_accuracy_all_v1_1 <- test_array_options_clean_v1_1 %>%
  group_by(participant,number_training_images,round,trial_counter,trial_type,current_category_training_level,current_category_label_level,current_category_kind,final_choice_array,total_number_correct_options_label,total_number_correct_options_training) %>%
  summarize(
    accuracy_training = mean(is_match_to_training, na.rm=TRUE),
    accuracy_label = mean(is_match_to_label,na.rm=TRUE),
    hit_rate_training = mean(hit_training,na.rm=TRUE),
    hit_rate_label = mean(hit_label,na.rm=TRUE),
    false_alarm_rate_training = mean(false_alarm_training,na.rm=TRUE),
    false_alarm_rate_label = mean(false_alarm_label,na.rm=TRUE)
  ) %>%
  mutate(
    hit_rate_training_adj= case_when(
      hit_rate_training==1 ~ 1 - 1/(2*total_number_correct_options_training),
      hit_rate_training==0 ~ 1/(2*total_number_correct_options_training),
      TRUE ~ hit_rate_training
    ),
    hit_rate_label_adj= case_when(
      hit_rate_label==1 ~ 1 - 1/(2*total_number_correct_options_label),
      hit_rate_label==0 ~ 1/(2*total_number_correct_options_label),
      TRUE ~ hit_rate_label
    ),
    false_alarm_rate_training_adj= case_when(
      false_alarm_rate_training==0 ~ 1/(2*total_number_correct_options_training),
      false_alarm_rate_training==1 ~ 1 - 1/(2*total_number_correct_options_training),
      TRUE ~ false_alarm_rate_training
    ),
    false_alarm_rate_label_adj= case_when(
      false_alarm_rate_label==0 ~ 1/(2*total_number_correct_options_label),
      false_alarm_rate_label==1 ~ 1 - 1/(2*total_number_correct_options_label),
      TRUE ~ false_alarm_rate_label
    )
  ) %>%
  mutate(
    dprime_training=qnorm(hit_rate_training_adj) - qnorm(false_alarm_rate_training_adj),
    dprime_label=qnorm(hit_rate_label_adj) - qnorm(false_alarm_rate_label_adj),
    c_training=-.5*(qnorm(hit_rate_training_adj) + qnorm(false_alarm_rate_training_adj)),
    c_label=-.5*(qnorm(hit_rate_label_adj) + qnorm(false_alarm_rate_label_adj)))

subj_test_accuracy_all_v1 <- full_join(sampling_data_v1,subj_test_accuracy_all_v1)

subj_test_accuracy_all_v1_1 <- full_join(sampling_data_v1_1,subj_test_accuracy_all_v1_1)

# gives accuracy based on training level, ground-truth level, and if sampling was confirm/explore

overall_accuracy_label_level_sampling_all_v1 <- subj_test_accuracy_all_v1 %>%
  ungroup() %>%
  group_by(sampling_choice_type,current_category_training_level,current_category_label_level) %>%
  summarize(
    N=n(),
    average_dprime_training=mean(dprime_training),
    average_dprime_label=mean(dprime_label)
  ) 

overall_accuracy_label_level_sampling_all_v1_1 <- subj_test_accuracy_all_v1_1 %>%
  ungroup() %>%
  group_by(sampling_choice_type,current_category_training_level,current_category_label_level) %>%
  summarize(
    N=n(),
    average_dprime_training=mean(dprime_training),
    average_dprime_label=mean(dprime_label)
  ) 
```

## Sampling

### Models

#### H1

Do learners flexibly shift their sampling choices depending on the training condition? 

```{r}
sampling_data_v1 <- sampling_data_v1 %>%
  mutate(
    sampled_category_level_kind_info_choice_order = case_when(
      sampled_category_level_kind_info == "outside_category" ~ "a_outside_category",
      sampled_category_level_kind_info == "sub" ~ "b_sub",
      sampled_category_level_kind_info == "bas" ~ "c_bas",
      sampled_category_level_kind_info == "sup" ~ "d_sup"
  ),
  current_category_training_level_order = case_when(
      current_category_training_level == "narrow" ~ "a_narrow",
      current_category_training_level == "intermediate" ~ "b_intermediate",
      current_category_training_level == "broad" ~ "c_broad"
  )
  )

sampling_data_v1_1 <- sampling_data_v1_1 %>%
  mutate(
    sampled_category_level_kind_info_choice_order = case_when(
      sampled_category_level_kind_info == "outside_category" ~ "a_outside_category",
      sampled_category_level_kind_info == "sub" ~ "b_sub",
      sampled_category_level_kind_info == "bas" ~ "c_bas",
      sampled_category_level_kind_info == "sup" ~ "d_sup"
  ),
  current_category_training_level_order = case_when(
      current_category_training_level == "narrow" ~ "a_narrow",
      current_category_training_level == "intermediate" ~ "b_intermediate",
      current_category_training_level == "broad" ~ "c_broad"
  )
  )

sampling_data_3_images_only_v1 <- sampling_data_v1 %>%
  filter(number_training_images==3)

sampling_data_3_images_only_v1_1 <- sampling_data_v1_1 %>%
  filter(number_training_images==3)

#fit multinomial model
nnetFixedModelStr <- 'sampled_category_level_kind_info_choice_order ~ current_category_training_level_order'

nnetFixedFit_v1 <-  multinom(as.formula(nnetFixedModelStr), sampling_data_3_images_only_v1, maxit=10000, abstol=1e-8, reltol=1e-10)
summary(nnetFixedFit_v1)
Anova(nnetFixedFit_v1,type="III")

nnetFixedFit_v1_1 <-  multinom(as.formula(nnetFixedModelStr), sampling_data_3_images_only_v1_1, maxit=10000, abstol=1e-8, reltol=1e-10)
summary(nnetFixedFit_v1_1)
Anova(nnetFixedFit_v1_1,type="III")

# Alternative (equivalent) approach with mlogit
# reformat data for mlogit
mlogit_sampling_data <-  mlogit.data(data=select(sampling_data_3_images_only,participant,sampled_category_level_kind_info_choice_order,current_category_training_level_order,current_category_kind), choice='sampled_category_level_kind_info_choice_order', shape="wide", id.var='participant')

# specify model formula
mlogit_sampling_formula_baseline <- ' sampled_category_level_kind_info_choice_order ~ 0 | 1 | 0'
mlogit_sampling_formula_training <- 'sampled_category_level_kind_info_choice_order ~ 0 | current_category_training_level_order | 0'

# fit models
mlogit_sampling_baseline <- mlogit(as.formula(mlogit_sampling_formula_baseline), data=mlogit_sampling_data, panel=FALSE, iterlim=10000, tol = 1e-8)
print(summary(mlogit_sampling_baseline))
mlogit_sampling_training <- mlogit(as.formula(mlogit_sampling_formula_training), data=mlogit_sampling_data, panel=FALSE, iterlim=10000, tol = 1e-8)
print(summary(mlogit_sampling_training))

# overall test of condition (identical result to nnet method above)
lrtest(mlogit_sampling_baseline,mlogit_sampling_training)
```

#### H1 across age

```{r}
#fit multinomial model
nnetFixedModelStr <- 'sampled_category_level_kind_info_choice_order ~ current_category_training_level_order*age_years'
nnetFixedFit <-  multinom(as.formula(nnetFixedModelStr), sampling_data_3_images_only, maxit=10000, abstol=1e-8, reltol=1e-10)
summary(nnetFixedFit)
Anova(nnetFixedFit,type="III")
```

#### 1-Narrow vs. 3-Narrow

```{r}
#filter relevant conditions
sampling_data_1_vs_3 <- sampling_data %>%
  filter(current_category_training_level=="narrow") %>%
  mutate(
    number_training_images_c=(number_training_images-2)/2
  )

#fit multinomial model
nnetFixedModelStr <- 'sampled_category_level_kind_info_choice_order ~ number_training_images_c'
nnetFixedFit <-  multinom(as.formula(nnetFixedModelStr), sampling_data_1_vs_3, maxit=10000, abstol=1e-8, reltol=1e-10)
summary(nnetFixedFit)
Anova(nnetFixedFit,type="III")
```

#### 1-Narrow vs. 3-Narrow across age

```{r}
#fit multinomial model
nnetFixedModelStr <- 'sampled_category_level_kind_info_choice_order ~ number_training_images_c*age_years'
nnetFixedFit <-  multinom(as.formula(nnetFixedModelStr), sampling_data_1_vs_3, maxit=10000, abstol=1e-8, reltol=1e-10)
summary(nnetFixedFit)
Anova(nnetFixedFit,type="III")
```


### Robustness Checks

#### Controlling for Category Type

```{r}
nnetFixedModelStr <- 'sampled_category_level_kind_info_choice_order ~ current_category_training_level_order+current_category_kind'
nnetFixedFit <-  multinom(as.formula(nnetFixedModelStr), sampling_data_3_images_only, maxit=10000, abstol=1e-8, reltol=1e-10)
summary(nnetFixedFit)
Anova(nnetFixedFit,type="II")
```

#### Adding by-participant random effects

```{r}
#set up random effects and model fitting parameters
parNames=names(mlogit_sampling_training$coefficients)[1:3]
print(parNames)
rparArg=rep('n',length(parNames))
names(rparArg)=parNames
print(rparArg)
NRCRDRAWS=1000

#fit models
if (NRCRDRAWS>0){
mlogit_sampling_baseline_re=mlogit(as.formula(mlogit_sampling_baseline), data=mlogit_sampling_data, rpar=rparArg, 
                panel=TRUE, R=NRCRDRAWS,print.level=1)
print(summary(mlogit_sampling_baseline_re))
}
if (NRCRDRAWS>0){
mlogit_sampling_training_re=mlogit(as.formula(mlogit_sampling_formula_training), data=mlogit_sampling_data, rpar=rparArg, 
                panel=TRUE, R=NRCRDRAWS,print.level=1)
print(summary(mlogit_sampling_training_re))
}

# LR Test
lrtest(mlogit_sampling_baseline_re,mlogit_sampling_training_re)
```

### Plot

```{r}
sampling_data$sampled_category_level_kind_info_ord <- factor(sampling_data$sampled_category_level_kind_info,
                                                             levels=c("sub","bas","sup","outside_category"),
                                                             labels=c("subordinate","basic","superordinate","outside category"))
sampling_data$current_category_training_level_ord <- factor(sampling_data$current_category_training_level,levels=c("narrow","intermediate","broad"))

ggplot(sampling_data,aes(current_category_training_level_ord,fill=sampled_category_level_kind_info_ord))+
  geom_bar(position="fill")+
  scale_fill_brewer(name="Sampling Choice Type",palette="RdYlBu")+
  xlab("Training Condition")+
  ylab("Proportion of Sampling Choices")+
  facet_wrap(~number_training_images)#+
  # theme_cowplot(font_size=24)+
  # theme(axis.title = element_text(face="bold", size=20),
  #       axis.text.x  = element_text(angle=90, hjust=1,vjust=0.4))

## Sampling Choices - individual bars
sampling_data_long_structure <- expand.grid(
  participant=unique(sampling_data$participant),
  trial_counter=seq(1,6),
  sampling_choice_type = c("subordinate","basic","superordinate","outside category")
)
sampling_data_long <-  sampling_data %>%
  group_by(participant,age_group,age_years,age_years_c,number_training_images,trial_counter, current_category_training_level) %>%
  summarize(sampling_response = sampled_category_level_kind_info_ord) %>%
  arrange(participant,trial_counter)
sampling_data_long_structure <- sampling_data_long_structure %>%
  left_join(sampling_data_long) %>%
  mutate(
    sampling_choice_chosen = case_when(
      sampling_response == sampling_choice_type ~ 1,
      TRUE ~ 0
    )
  )

sampling_data_long_structure_summarized <- sampling_data_long_structure %>%
  group_by(number_training_images,current_category_training_level, sampling_choice_type) %>%
  summarize(
    N=n(),
    sampling_response_avg=mean(sampling_choice_chosen,na.rm=TRUE),
    sampling_response_total=sum(sampling_choice_chosen,na.rm=TRUE),
    sampling_response_binom_lower=binom.test(sampling_response_total,N)$conf.int[1],
    sampling_response_binom_upper=binom.test(sampling_response_total,N)$conf.int[2],
  )
sampling_data_long_structure_summarized$current_category_training_level_ord <- factor(sampling_data_long_structure_summarized$current_category_training_level,levels=c("narrow","intermediate","broad"))


ggplot(sampling_data_long_structure_summarized,aes(current_category_training_level_ord,sampling_response_avg, group=sampling_choice_type,fill=sampling_choice_type))+
  geom_errorbar(aes(ymin=sampling_response_binom_lower, ymax=sampling_response_binom_upper),width=0.3, color="black",position=position_dodge(width=0.9))+
  geom_bar(stat="identity",position=position_dodge())+
  #theme_cowplot(font_size=30)+
  scale_fill_brewer(name="Sampling Choice Type",palette="RdYlBu")+
  scale_color_brewer(name="Sampling Choice Type",palette="RdYlBu")+
  xlab("Training Condition")+
  ylab("Proportion of Sampling Choices")+
  facet_wrap(~number_training_images)
```

## Test

### Plot

```{r}
## Test Plot
subj_test_prop <- test_array_options_clean %>%
  mutate(training_condition = case_when(
    number_training_images == 1 ~ "1-Item",
    TRUE ~ paste0("3-Item ", stringr::str_to_title(current_category_training_level))
  )) %>%
  ungroup() %>%
  group_by(participant,age_years,age_group,age_years_c,number_training_images,current_category_training_level,training_condition,current_category_label_level,current_category_kind) %>%
  summarize(
    N_subordinate = sum(test_image_type=="subordinate" & test_image_match_category==1),
    prop_subordinate = sum(test_image_type=="subordinate" & test_image_match_category==1 & is_chosen == 1)/sum(test_image_type=="subordinate" & test_image_match_category==1),
    N_basic = sum(test_image_type=="basic" & test_image_match_category==1),
    prop_basic = sum(test_image_type=="basic" & test_image_match_category==1 & is_chosen == 1)/sum(test_image_type=="basic" & test_image_match_category==1),
    N_superordinate = sum(test_image_type=="superordinate" & test_image_match_category==1),
    prop_superordinate = sum(test_image_type=="superordinate" & test_image_match_category==1 & is_chosen == 1)/sum(test_image_type=="superordinate" & test_image_match_category==1)
  )

subj_test_prop_long <- subj_test_prop %>%
  pivot_longer(N_subordinate:prop_superordinate,names_to=c(".value","choice_type"),names_sep="\\_")

overall_test_prop <- subj_test_prop %>%
  group_by(training_condition,number_training_images,current_category_training_level) %>%
  summarize(
    N=n(),
    subordinate_percent = mean(prop_subordinate),
    basic_percent = mean(prop_basic),
    superordinate_percent = mean(prop_superordinate)
  )

overall_test_prop_long <- subj_test_prop_long %>%
  group_by(training_condition,number_training_images,current_category_training_level,choice_type) %>%
  summarize(
    N=n(),
    avg_prop = mean(prop),
    prop_ci=qt(0.975, N-1)*sd(prop,na.rm=T)/sqrt(N),
    prop_lower_ci=avg_prop-prop_ci,
    prop_upper_ci=avg_prop+prop_ci,
  )
overall_test_prop_long$choice_type_ord <- factor(overall_test_prop_long$choice_type,levels=c("subordinate","basic","superordinate"))
overall_test_prop_long$training_condition_ord <- factor(overall_test_prop_long$training_condition,levels=c("1-Item","3-Item Narrow", "3-Item Intermediate","3-Item Broad"))

ggplot(overall_test_prop_long,aes(choice_type_ord,avg_prop,fill=choice_type_ord))+
  geom_bar(stat="identity")+
  geom_errorbar(aes(ymin=prop_lower_ci,ymax=prop_upper_ci),width=0.1)+
  facet_wrap(~training_condition_ord)+
  theme_cowplot(font_size=12)+
  scale_fill_manual(
    name="Test Choice Type",
    values=c("#d7191c","#fdae61","#abd9e9"))+
  xlab("Test Choice Type")+
  ylab("Average Proportion\nof Test Choices")+
  theme(legend.position="none")+
  theme(axis.title = element_text(face="bold", size=14),
        #axis.text.x  = element_text(angle=90, hjust=1,vjust=0.4)
        )
ggsave(here(figure_path,"kitty-act-test.png"),width=6,height=6)
  
```

## Sampling by Age

### Plot

```{r fig.width=10, fig.height=4}

# summarizing
sampling_data_ages_summarized <- sampling_data_long_structure %>%
  group_by(age_group, 
           number_training_images,
           current_category_training_level,
           sampling_choice_type) %>% 
  summarize(
    N=n(),
    sampling_response_avg=mean(sampling_choice_chosen,na.rm=TRUE),
    sampling_response_total=sum(sampling_choice_chosen,na.rm=TRUE),
    sampling_response_binom_lower=binom.test(sampling_response_total,N)$conf.int[1],
    sampling_response_binom_upper=binom.test(sampling_response_total,N)$conf.int[2],
  )

sampling_data_ages_summarized$current_category_training_level <- factor(sampling_data_ages_summarized$current_category_training_level,levels=c("narrow","intermediate","broad"))

sampling_data_ages_summarized %>% filter(number_training_images == 1) %>% 
  ggplot(aes(x = age_group, 
             y = sampling_response_avg, 
             color = sampling_choice_type)) + 
  geom_point() +
  geom_line()

sampling_data_ages_summarized %>% filter(number_training_images == 3) %>% 
  ggplot(aes(x = age_group, 
             y = sampling_response_avg, 
             color = sampling_choice_type)) + 
  geom_point() +
  geom_line() + 
  facet_wrap(. ~ current_category_training_level, ncol = 3)
```

Continuous plot

```{r}
ggplot(sampling_data_long_structure,aes(age_years,sampling_choice_chosen,color=sampling_choice_type))+
  geom_smooth(method="loess")+
  facet_wrap(~number_training_images+current_category_training_level)
```

## Exploratory

### investigate the proportion of exploratory choices

```{r}
sampling_data <- sampling_data %>%
  mutate(is_exploratory=ifelse(sampling_choice_type=="explore",1,0)) %>%
  mutate(training_condition = case_when(
    number_training_images == 1 ~ "1-Item",
    TRUE ~ paste0("3-Item ", stringr::str_to_title(current_category_training_level))
  )) 

subj_exploratory <- sampling_data %>%
  mutate(
    age_bins=case_when(
      age_years<5.5 ~ 5,
      age_years<6.5 ~ 6,
      age_years<7.5 ~ 7,
      age_years<8.5 ~ 8,
      age_years<9.5 ~ 9,
      TRUE ~ NA
      
    )
  ) %>%
  group_by(participant,age_bins) %>%
  summarize(
    N=n(),
    prop_exploratory=mean(is_exploratory)
  )

#exploratory averages by age group
overall_exploratory_age_group <- subj_exploratory %>%
  group_by(age_bins) %>%
  summarize(
    N=n(),
    mean_explore=mean(prop_exploratory),
    sd(prop_exploratory,na.rm=T)/sqrt(N),
    se = sd(prop_exploratory,na.rm=T)/sqrt(N),
    ci=qt(0.975, N-1)*se,
    lower_se = mean_explore-se,
    upper_se = mean_explore+se,
    lower_ci=mean_explore-ci,
    upper_ci=mean_explore+ci,
  )

ggplot(sampling_data,aes(age_years,is_exploratory))+
  geom_hline(yintercept=0.5,linetype="dashed")+
  geom_smooth(method="lm",color="black")+
  geom_errorbar(data=overall_exploratory_age_group,aes(age_bins,y=mean_explore,ymin=lower_se,ymax=upper_se),width=0,color="#0A5C36")+
  geom_point(data=overall_exploratory_age_group,aes(age_bins,mean_explore),color="#0A5C36",size=2)+
  ylab("Proportion of Exploratory Sampling Choices")+
  xlab("Age (in years)")+
  theme_cowplot(font_size=12)+
  theme(axis.title = element_text(face="bold", size=14),
        #axis.text.x  = element_text(angle=90, hjust=1,vjust=0.4)
        )
ggsave(here(figure_path,"kitty-act-sampling-by-age.png"),width=6,height=6)
 

ggplot(sampling_data,aes(age_years,is_exploratory, color=training_condition))+
  geom_smooth(method="lm",se=F)+
  geom_smooth(method="lm",color="black")

#model
m <- glmer(is_exploratory~ age_years_c*training_condition+(1|participant),data=sampling_data,family="binomial")
summary(m)
Anova(m,type="III")
  
```

## H4: Test Choices

## Training condition modulated participants’ choices at test (Proportion choices)

```{r}
# fit individual lmer models
## subordinate choices
m_sub <- lmer(prop_subordinate~training_condition+(1|participant),data=subj_test_prop)
summary(m_sub)
Anova(m_sub,type="III")

## basic choices
m_bas <- lmer(prop_basic~training_condition+(1|participant),data=subj_test_prop)
summary(m_bas)
Anova(m_bas,type="III")

## superordinate choices
m_sup <- lmer(prop_superordinate~training_condition+(1|participant),data=subj_test_prop)
summary(m_sup)
Anova(m_sup,type="III")

```

### H5: Proportion basic-level choices

#### Test

```{r}
subj_test_prop <- subj_test_prop %>%
  mutate(number_training_images_c=(number_training_images-2)/2)

m <- glmer(prop_basic ~ number_training_images_c+(1|participant)+(1|current_category_kind),data=filter(subj_test_prop,current_category_training_level=="narrow"),family="binomial")
summary(m)
#replicates the basic Xu/ Tenenbaum effect
```

Does this interact with Age?

```{r}
m <- glmer(prop_basic ~ number_training_images_c*age_years_c+(1|participant)+(1|current_category_kind),data=filter(subj_test_prop,current_category_training_level=="narrow"),family="binomial")
summary(m)
```
